+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Retrieve missing publishers from Crossref
 - Add column indicating original research (as opposed to review, synthesis, opinion, etc)
 - Retrieve missing affiliations from OpenAlex (these are very messy currently)
+- Extract missing countries from titles and abstracts (best effort, still many missing)
 
 ## [1.0.0] - 2024-12-09